Fix flow-of-control error in geonode.maps.utils.upload
This method has three code paths - one for individual files, one for
directories, and another for reporting to the user that he invoked the
method incorrectly.  Prior to this commit, invoking the method with a
single file would invoke both the single-file and directory code paths,
and promptly blow up since a file is not a directory.