Corrections to the way geonode.maps.utils talks to GeoNetwork and GeoServer
There seemed to be a misconception about gsconfig's
FailedRequestException - it indicates that GeoServer has failed to
respond (either by refusing the network connection or by giving back
something other than the XML gsconfig expects.)  When an item does not
exist in GeoServer, gsconfig responds with None rather than raising.

The code to delete records from GeoNetwork was also invalid - it
expected a method on the geonode.geonetwork.Catalog object that simply
isn't there.  It would probably be nice to be able to delete based on
just a UUID but for now I put a small hack in, passing a dict() rather
than the Layer model instance that Catalog.delete_layer() was written
for.

Finally, I added a call to geonode.geonetwork.Catalog.login to the code
that checks whether GeoServer and Django are up.  I had originally been
suspicious of this code not checking whether Django is up but when I
thought about it, it doesn't really need to. Since the Django app itself
doesn't need to make HTTP requests to the Django app, it's enough to
verify that GeoServer is responding (which does indirectly require
Django to be listening on HTTP.)