Deal with cross lib dependencies.
After the recent GeoExt upgrade, we need to be able to exclude directories from builds. This functionality is provided by recent versions of buildkit, but pulling this in also requires modifications to autoloader.js. This change is basically the latest from readygxp, with GeoNode specific modifications for the pathMappings from the old config.js. These are now in main.js, and are dealt with by a convenient 301 redirect.