Clean up the global search javascript namespace
Make all of the code pass jslint. This was mostly spacing issues like
a space after the function keyword and before the ()

Fix the casing on the main Class used in the search module.

Force of the use of square brackets for if else statements. While it
is legal not to use square brackets on single line if else statements,
that code becomes hard to maintain.

Remove use of global vars. The old code used global vars that polluted
the window namespace. The new method requires that you pass in the
keywords and categories vars.