Add missing layers' sources in map composer
When viewing an existing map (GeoNode map composer), only the sources for  the layers included in the map are available in the "View available data from" dropdown menu ("Add layers" tool). The complete list of sources (defined in MAP_BASELAYERS) is not available any more.

I think It is desirable to always add the missing layers' sources to dropdown menu. At least, it seems that many of our users have this need.

It is true that a missing WMS endpoint can be added manually, but can be quite hard to recover the correct URL.
Moreover,   layers of different type (eg. bing, mapquest) can no longer be added.

Here a simple patch to check the missing sources and add them to the map's configuration.