Change to PRINT_NG, check for PRINTNG
PRINTNG_ENABLED is obscure and looks like a typo.
Changing it to PRINTING_ENABLED doesn't work either, because it exists
in distinction to MAPFISH_PRINT_ENABLED, which is still in use.
So change to PRINT_NG_ENABLED.

If PRINTNG_ENABLED is encountered in config, use this, and log a
deprecation message.