catch and fix import error
revisiting a previous issue.  The problem has been isolated and
in contrast with a previous PR, the problem statement and
fix in a try catch block has been isolated to where the problem
occurs as opposed to a top level, beginning of file  sys.path.append

The ALLOWED_HOSTS setting needs to be a list but
there is no way to make it so just by using os.getenv

Added logic to check if the environment variable exits
and if so, split on comma. colon, or semicolon excluding
leading or trailing spaces