Update workspace buttons layout and text
Previously the "Set permissions" and "Create a
Map" buttons were designed in such a way that the
text would overflow the buttons and collide with
each other at certain page widths. This behavior
isn't desired, and the buttons were redesigned to
have a clean look and feel regardless of the page
width.