Implement aborting of harvesting celery tasks
Add the `AsynchronousHarvestingSession` model, which is used
to implement sessions for both refreshing of a harvester's
harvestable resources and for the harvesting of remote
resources.

Refactor the `admin`, `api` and `tasks` to use this new model.

Moved some functions out of `harvesting.utils` module in order
to avoid circular imports