[Fixes #13049] Update importlayers command to handle all data formats supported by the upload API
A refactor of the importlayers script so it supports all data formats supported by the upload API. It makes use of the same BaseHandler classes used by the importer to check whether a file is valid for upload and whether all of the required files are present (it also checks for optional files).

It also adds the --overwrite-existing-layers and --skip-existing-layers flags. Although the --skip-existing-layers flag doesn't appear to be supported in the upload API yet.

By using the same BaseHandler classes as the importer the aim is help keep the importlayers command in sync with the importer. For example, if additional file extensions are added to handlers, or a completely new handler is introduced, the importlayers command should start supporting it without any additional changes required.

It treats ".zip" and ".kmz" files differently as these do not have a Handler class and instead get unpacked by the StorageManager after being sent to the upload API.